Organized data models into separate files
