+++ models/base.py
@@ -1,0 +1,8 @@
+"""
+Base database setup and common imports for models
+"""
+from datetime import datetime, timezone
+from flask_sqlalchemy import SQLAlchemy
+
+# We'll get db from the app context instead of importing directly
+db = SQLAlchemy()
